Navigation and Planner Fixes
Adjusted how Pawns enter and exit furniture, and fixed a bug with ChairSprites interaction points list. Added the IPlanTask for Tasks that don't really need to be used by Pawns but are just there for housekeeping purposes with the Planner. Currently StanceStand is the only IPlanTask, but it may be necessary in the future.